requirements.txt
asgiref==3.7.2
black==22.1.0
blue==0.9.1
click==8.1.7
Django==5.0.2
djangorestframework==3.14.0
flake8==4.0.1
mccabe==0.6.1
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.2.0
pycodestyle==2.8.0
pyflakes==2.4.0
pytz==2024.1
sqlparse==0.4.4
tomli==2.0.1
